edits electronics.java

+++ Electronics.java
@@ -6,7 +6,7 @@ public class Electronics extends Product {
     }
 
 
-    public Electronics(String productID, String productName, int noOfAvailableItems, int price, String brand, int warrantyPeriod){
+    public Electronics(String productID, String productName, int noOfAvailableItems, double price, String brand, int warrantyPeriod){
         super(productID, productName, noOfAvailableItems, price);
         this.brand=brand;
         this.warrantyPeriod=warrantyPeriod;
